target-arm: handle CONSTRAINED UNPREDICTABLE ldp/stxr
Ed reported a bug in OpenJDK's register allocation that led to code
using ldp and stxr working on QEMU but not on real hardware. Closer
reading of the ARM ARM implies several implementation options are
possible but generating an UNDEFINED instruction error seems the most
sane.

(COMPILE TESTED ONLY)

Reported-by: Edward Nevill <edward.nevill@linaro.org>
Signed-off-by: Alex Bennée <alex.bennee@linaro.org>

---

v1:
  - RFC until after KVM Forum and I can properly test this.