configure: move EXTRA_CFLAGS append to the end
When using --extra-cflags to override defaults the flags need to be set
at the end lest they be reset by later options. This affects
optimisation as well where "-O0 .. -O3" will just takes the most recent
option.

We also set CFLAGS so the options are passed to other built binaries.

Signed-off-by: Alex Bennée <alex.bennee@linaro.org>