include/qemu/atomic.h: use atomic builtins when sanitizing
When running under ThreadSanitizer the existing atomic_read() is not
good enough as it gets treated as a potentially non-atomic read. By
explicitly using the __atomic_load primitive we suppress a bunch of false
positive warnings. The same is true for atomic_set which will use the
__atomic_store builtin.

These changes are gated by the __USING_TSAN define which the configure
script adds when building with -fsantize=thread.

Signed-off-by: Alex Bennée <alex.bennee@linaro.org>