async: use atomic_read() instead of plain access
Both bh->scheduled and ctx->notify_me have been set with atomic_*
primitives so ThreadSanitizer complains when it sees them accessed
without atomics. While this probably works in most cases (at least on
x86) the compiler could pull tricks if not constrained.

In a non-sanitizer build on x86 the atomic_read at least ensures a
volatile is applied to the value.

Signed-off-by: Alex Bennée <alex.bennee@linaro.org>