exec: make phys_page_compact_all safely handle nodes_nb == 0
when compiling with -fsanitize=undefined the compiler complains for a
potential negative or zero length array definition from DECLARE_BITMAP.
Although unlikely this would lead to the stack being smashed.

As nodes_nb is passed from PhysPageMap nodes_nb where it is unsigned we
fix the prototype to accept an unsigned count. We also guard the
declaration inside an if (nodes_nb > 0) leg.

Signed-off-by: Alex Bennée <alex.bennee@linaro.org>